BJS2-56387 1. Добавленна папка с названием задачи для файлов используемых в задаче. 2. Добавленна пустая строка между блоками констант и полей класса. 3. Добавленна проверка максимальной длины имени. 4. В сообщения об ошибках добавленно текущее значение параметра. 5. Логика валидации вынесена в отдельные методы.
