+++ src/main/java/school/faang/BJS2_56387/Main.java
@@ -1,0 +1,4 @@
+package school.faang.BJS2_56387;
+
+public class Main {
+}
