--- src/main/java/school/faang/Main.java
@@ -1,4 +1,0 @@
-package school.faang;
-
-public class Main {
-}
